Better Storybook configuration
